requirements.txt
importd==0.4.3
django-storages==1.1.8
boto==2.38.0
celery==3.1.19
waitress==0.8.9
python-binary-memcached
django-bmemcached
